# ENH: big performance boost for first visualization