All types in Vega.jl are now subtypes of `AbstractVegaType`. This allowed
us to do some clever performance tricks...

The tricks come in two parts:

1. Instead of defining `TypeName(;all_fields=their_defaults...)` we now just
define `TypeName() = TypeName(default1, default2...)`
2. Define `Base.call{T<:AbstractVegaType}(::Type{T}; kwargs...)` to first call
that no argument constructor to fill in default avlues and then fill in
additional values users pass as keyword arguments.

Nothing at all changes about how we interact with the types. The only file that
needed adjusting to accomodate this change is `src/jstypes.jl`

Here is an idea of the performance boost we get:

```
Vega|sl/speedy⚡ ⇒ julia
               _
   _       _ _(_)_     |  A fresh approach to technical computing
  (_)     | (_) (_)    |  Documentation: http://docs.julialang.org
   _ _   _| |_  __ _   |  Type "help()" for help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 0.4.0-dev+6176 (2015-07-22 16:13 UTC)
 _/ |\__'_|_|_|\__'_|  |  Commit 370af84 (0 days old master)
|__/                   |  x86_64-apple-darwin14.4.0

julia> include(Pkg.dir("Vega", "src", "Vega.jl"));

julia> tic(); lp = Vega.lineplot(x=1:10, y=sin(1:10)); toc();
elapsed time: 0.619851575 seconds

julia> tic(); lp = Vega.lineplot(x=1:10, y=sin(1:10)); toc();
elapsed time: 0.00069959 seconds

julia>
Vega|sl/speedy⚡ ⇒ git checkout vega2
Switched to branch 'vega2'
Your branch is up-to-date with 'fork/vega2'.
Vega|vega2⚡ ⇒ julia
               _
   _       _ _(_)_     |  A fresh approach to technical computing
  (_)     | (_) (_)    |  Documentation: http://docs.julialang.org
   _ _   _| |_  __ _   |  Type "help()" for help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 0.4.0-dev+6176 (2015-07-22 16:13 UTC)
 _/ |\__'_|_|_|\__'_|  |  Commit 370af84 (0 days old master)
|__/                   |  x86_64-apple-darwin14.4.0

julia> include(Pkg.dir("Vega", "src", "Vega.jl"));

julia> tic(); lp = Vega.lineplot(x=1:10, y=sin(1:10)); toc();
elapsed time: 38.781551306 seconds

julia> tic(); lp = Vega.lineplot(x=1:10, y=sin(1:10)); toc();
elapsed time: 0.000592659 seconds
```

That's an over 60x improvement for the first visualization with no performance penalty for subsequent calls!