@randyzwitch 2015-07-23
Wait, what? Hahaha.

I literally have no idea what's going on here, so I'll just take your word for it. This is some voodoo that I just can't even wrap my head around right now.