@spencerlyon2 2015-07-23
Haha it is definitely some voodoo magic that I don't fully understand, but here's what I _think_ happened (sorry, I'm writing this out mostly to convince myself I have some idea of what happened -- you just get to be my audience :smirk:)

Before we had keyword argument constructors for every type that listed every field with its default value. I think this resulted in many internal (we don't see or call them ourselves) positional constructors being defined/compiled based on what combination of 

In this case I think we only get two constructors per type `T`:

1. The `T() = T(default1, default2...)`
2. `Base.call(::Type{T}; kwargs::Vector{Tuple{Symbol, Any}})`
    - We get this one through "inheritance"  because it is parametric over all `AbstractVegaTypes`
    - I think this is way of doing keyword arguments is more efficient in our case than before because it doesn't need to compile a new method based on which combination of keyword arguments were actually passed. We effectively just gave up on that and told the compiler "you just collect them in a vector like `Tuple{Symbol, Any}[(:passed_arg_1, value1), (:passed_arg_2, value2), ...]` and I'll deal with them myself".
    - This should be equivalent to what we had before because users can pass any keyword/value they want and if the keyword is invalid we see `ERROR: type XXXX has no field ZZZZ` instead of `NoMethodError` and if the type of the value is wrong we see `ERROR: TypeError: setfield!: expected RIGHT_TYPE, got WRONG_TYPE` instead of `NoMethodError`.

Hopefully that made some sense and wasn't just flat out wrong...